+++ pioreactor/automations/temperature/__init__.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 
 from .constant_duty_cycle import ConstantDutyCycle
 from .only_record_ambient_temperature import OnlyRecordAmbientTemperature
-from .stable import Stable
+from .thermostat import Thermostat
 
 
-__all__ = ("OnlyRecordAmbientTemperature", "ConstantDutyCycle", "Stable")
+__all__ = ("OnlyRecordAmbientTemperature", "ConstantDutyCycle", "Thermostat")